refactor(dma): split into register and structures
